Load Lists and List items from the backend
This change fills out the AllListsView and the ListDetailsView pages by
loading the content from the backend.

This also saves the authentication token to be used for API calls and retrives
it using the UseAuth hook.

The client for the API is introduced in YataClient.js. The client is rather
basic right now and just wraps the fetch api calls from the backend. The client
can be extensively further configured in the future in a refactoring once
we understand the use cases for the client better.