Add ability to create new lists
This change adds a form to the AllListsView for a user to create new lists.

A new form component is defined in CreateListView. When a new list is created,
it is added to the view for the user to interact with immediately.

It also pushes the duty to create sections and containers for formatting to
lower components instead of App, for each view to have better control over
its layer.